# DSell_API_OpenAPI.json
{
    "openapi":  "3.0.3",
    "info":  {
                 "title":  "DSell 直销管理系统 API",
                 "version":  "1.0.0",
                 "description":  "Base URL: http://localhost:8080\n认证方式: Header 中携带 token (登录接口返回)"
             },
    "servers":  [
                    {
                        "url":  "http://localhost:8080"
                    }
                ],
    "security":  [
                     {
                         "token":  [

                                   ]
                     }
                 ],
    "components":  {
                       "securitySchemes":  {
                                               "token":  {
                                                             "type":  "apiKey",
                                                             "in":  "header",
                                                             "name":  "token"
                                                         }
                                           },
                       "schemas":  {
...
                                       "ProductDTO":  {
                                                          "type":  "object",
                                                          "required":  [
                                                                           "name"
                                                                       ],
                                                          "properties":  {
                                                                             "id":  {
                                                                                        "type":  "integer"
                                                                                    },
                                                                             "name":  {
                                                                                          "type":  "string"
                                                                                      },
                                                                             "categoryId":  {
                                                                                                "type":  "integer"
                                                                                            },
                                                                             "subtitle":  {
                                                                                              "type":  "string"
                                                                                          },
                                                                             "mainImage":  {
                                                                                               "type":  "string"
                                                                                           },
                                                                             "images":  {
                                                                                            "type":  "string"
                                                                                        },
                                                                             "detail":  {
                                                                                            "type":  "string"
                                                                                        },
                                                                             "retailPrice":  {
                                                                                                 "type":  "number"
                                                                                             },
                                                                             "memberPrice":  {
                                                                                                 "type":  "number"
                                                                                             },
                                                                             "pv":  {
                                                                                        "type":  "number"
                                                                                    },
                                                                             "stock":  {
                                                                                           "type":  "integer"
                                                                                       },
                                                                             "isRecommend":  {
                                                                                                 "type":  "integer",
                                                                                                 "enum":  [
                                                                                                              0,
                                                                                                              1
                                                                                                          ]
                                                                                             },
                                                                             "isNew":  {
                                                                                           "type":  "integer",
                                                                                           "enum":  [
                                                                                                        0,
                                                                                                        1
                                                                                                    ]
                                                                                       },
                                                                             "status":  {
                                                                                            "type":  "integer",
                                                                                            "enum":  [
                                                                                                         0,
                                                                                                         1
                                                                                                     ]
                                                                                        }
                                                                         }
                                                      },
...
                                   }
                   },
...
    "paths":  {
...
                  "/api/product/page":  {
                                            "get":  {
                                                        "tags":  [
                                                                     "Product"
                                                                 ],
                                                        "summary":  "分页查询产品",
                                                        "parameters":  [
                                                                           {
                                                                               "name":  "page",
                                                                               "in":  "query",
                                                                               "schema":  {
                                                                                              "type":  "integer",
                                                                                              "default":  1
                                                                                          }
                                                                           },
                                                                           {
                                                                               "name":  "size",
                                                                               "in":  "query",
                                                                               "schema":  {
                                                                                              "type":  "integer",
                                                                                              "default":  10
                                                                                          }
                                                                           },
                                                                           {
                                                                               "name":  "keyword",
                                                                               "in":  "query",
                                                                               "schema":  {
                                                                                              "type":  "string"
                                                                                          }
                                                                           },
                                                                           {
                                                                               "name":  "categoryId",
                                                                               "in":  "query",
                                                                               "schema":  {
                                                                                              "type":  "integer"
                                                                                          }
                                                                           },
                                                                           {
                                                                               "name":  "status",
                                                                               "in":  "query",
                                                                               "schema":  {
                                                                                              "type":  "integer"
                                                                                          }
                                                                           }
                                                                       ]
                                                    }
                                        },
                  "/api/product/{id}":  {
                                            "get":  {
                                                        "tags":  [
                                                                     "Product"
                                                                 ],
                                                        "summary":  "获取产品详情",
                                                        "parameters":  [
                                                                           {
                                                                               "name":  "id",
                                                                               "in":  "path",
                                                                               "required":  true,
                                                                               "schema":  {
                                                                                              "type":  "integer"
                                                                                          }
                                                                           }
                                                                       ]
                                                    }
                                        },
                  "/api/product":  {
                                       "post":  {
                                                    "tags":  [
                                                                 "Product"
                                                             ],
                                                    "summary":  "新增产品",
                                                    "requestBody":  {
                                                                        "content":  {
                                                                                        "application/json":  {
                                                                                                                 "schema":  {
                                                                                                                                "$ref":  "#/components/schemas/ProductDTO"
                                                                                                                            }
                                                                                                             }
                                                                                    }
                                                                    }
                                                },
                                       "put":  {
                                                   "tags":  [
                                                                "Product"
                                                            ],
                                                   "summary":  "更新产品",
                                                   "requestBody":  {
                                                                       "content":  {
                                                                                       "application/json":  {
                                                                                                                "schema":  {
                                                                                                                               "$ref":  "#/components/schemas/ProductDTO"
                                                                                                                           }
                                                                                                            }
                                                                                   }
                                                                   }
                                               }
                                   },
                  "/api/product/{id}/status":  {
                                                   "put":  {
                                                               "tags":  [
                                                                            "Product"
                                                                        ],
                                                               "summary":  "修改产品状态",
                                                               "parameters":  [
                                                                                  {
                                                                                      "name":  "id",
                                                                                      "in":  "path",
                                                                                      "required":  true,
                                                                                      "schema":  {
                                                                                                     "type":  "integer"
                                                                                                 }
                                                                                  },
                                                                                  {
                                                                                      "name":  "status",
                                                                                      "in":  "query",
                                                                                      "required":  true,
                                                                                      "schema":  {
                                                                                                     "type":  "integer",
                                                                                                     "enum":  [
                                                                                                                  0,
                                                                                                                  1
                                                                                                              ]
                                                                                                 }
                                                                                  }
                                                                              ]
                                                           }
                                               },
                  "/api/product/{id}/stock":  {
                                                  "put":  {
                                                              "tags":  [
                                                                           "Product"
                                                                       ],
                                                              "summary":  "修改库存",
                                                              "parameters":  [
                                                                                 {
                                                                                     "name":  "id",
                                                                                     "in":  "path",
                                                                                     "required":  true,
                                                                                     "schema":  {
                                                                                                    "type":  "integer"
                                                                                                }
                                                                                 },
                                                                                 {
                                                                                     "name":  "stock",
                                                                                     "in":  "query",
                                                                                     "required":  true,
                                                                                     "schema":  {
                                                                                                    "type":  "integer"
                                                                                                }
                                                                                 }
                                                                             ]
                                                          }
                                              },
...
                  "/api/product/category/{id}":  {
                                                     "get":  {
                                                                 "tags":  [
                                                                              "Product"
                                                                          ],
                                                                 "summary":  "获取分类详情",
                                                                 "security":  [

                                                                              ],
                                                                 "parameters":  [
                                                                                    {
                                                                                        "name":  "id",
                                                                                        "in":  "path",
                                                                                        "required":  true,
                                                                                        "schema":  {
                                                                                                       "type":  "integer"
                                                                                                   }
                                                                                    }
                                                                                ]
                                                             },
                                                     "delete":  {
                                                                    "tags":  [
                                                                                 "Product"
                                                                             ],
                                                                    "summary":  "删除分类",
                                                                    "security":  [

                                                                                 ],
                                                                    "parameters":  [
                                                                                       {
                                                                                           "name":  "id",
                                                                                           "in":  "path",
                                                                                           "required":  true,
                                                                                           "schema":  {
                                                                                                          "type":  "integer"
                                                                                                      }
                                                                                       }
                                                                                   ]
                                                                }
                                                 },
...
              }
}